Add a program to launch viewers in the background.
This is an initial spike at a bundled equivalent to the mutt_bgrun
script with some extensions.

In addition to launching an external viewer program with no text output,
mutt_launch should

- close open file desriptors inherited from the main mutt process.
- clean up temp files after the viewer process closes

Only the first of these is attempted in this version and I'm not even
sure that this is the best way to do it.

I haven't changed anything in mutt_view_attachment yet.