mutt_launch: Handle closing leaked file descriptors better.
On the advice of flatcap in #93, use `getrlimit` to get the actual maximum
number of open file descriptors.  Additionally, it now checks that the file
descriptor either closes succesfully or fails with EBADF (bad file descriptor
number) or fails and exits with a message to standard error about the leaked
file descriptor.  If this causes a large number of issues we might be able to
add a few retries before exiting but I'm hesitant to do so prematurely.